restructure project a little bit
