+++ first_flutter_app/lib/features/favorite/widgets/favorite_item.dart
