+++ first_flutter_app/lib/features/home/logic/app_state.dart
@@ -14,10 +14,10 @@ class AppState extends Equatable {
 
   @override
   List<Object?> get props => [
-    currentPair,
-    favorites,
-    history,
-  ];
+        currentPair,
+        favorites,
+        history,
+      ];
 
   AppState copyWith({
     WordPair? currentPair,
