+++ first_flutter_app/lib/features/home/widgets/word_gen_view.dart
@@ -1,7 +1,7 @@
 import 'package:english_words/english_words.dart';
-import 'package:first_flutter_app/bloc/app_cubit.dart';
-import 'package:first_flutter_app/home_widgets/history_words_viewed.dart';
-import 'package:first_flutter_app/home_widgets/word_card.dart';
+import 'package:first_flutter_app/features/home/logic/app_cubit.dart';
+import 'package:first_flutter_app/features/home/widgets/history_words_viewed.dart';
+import 'package:first_flutter_app/features/home/widgets/word_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -39,12 +39,13 @@ class WordGenView extends StatelessWidget {
                         context.read<AppCubit>().toggleFavorite(pair);
                       },
                       icon: Icon(
-                        state.favorites.contains(pair) ? Icons.favorite : Icons.favorite_border,
+                        state.favorites.contains(pair)
+                            ? Icons.favorite
+                            : Icons.favorite_border,
                       ),
                       label: const Text('Like'),
                     ),
                     const SizedBox(width: 10),
-
                     ElevatedButton(
                       onPressed: context.read<AppCubit>().getNext,
                       child: const Text('Next'),
