+++ first_flutter_app/lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:first_flutter_app/bloc/app_cubit.dart';
-import 'package:first_flutter_app/home_page.dart';
+import 'package:first_flutter_app/features/home/logic/app_cubit.dart';
+import 'package:first_flutter_app/features/home/view/home_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -21,7 +21,7 @@ class MyApp extends StatelessWidget {
           colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepOrange),
           useMaterial3: true,
         ),
-        home: const HomePage(),
+        home: const HomeView(),
       ),
     );
   }
